requirements/common.txt
Django==5.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==23.5
stripe==7.10.0
celery[redis]==5.3.6
flower==2.0.1